修改推荐积分功能

+++ admin/goods.php
@@ -818,9 +818,9 @@ elseif ($_REQUEST['act'] == 'insert' || $_REQUEST['act'] == 'update')
     $goods_number = isset($_POST['goods_number']) ? $_POST['goods_number'] : 0;
     $warn_number = isset($_POST['warn_number']) ? $_POST['warn_number'] : 0;
     $goods_type = isset($_POST['goods_type']) ? $_POST['goods_type'] : 0;
-    $give_integral = isset($_POST['give_integral']) ? intval($_POST['give_integral']) : '-1';
+    $give_integral = isset($_POST['give_integral']) ? floatval($_POST['give_integral']) : '-1';
 	$fencheng = isset($_POST['fencheng']) ? $_POST['fencheng'] : '0';
-    $rank_integral = isset($_POST['rank_integral']) ? intval($_POST['rank_integral']) : '-1';
+    $rank_integral = isset($_POST['rank_integral']) ? floatval($_POST['rank_integral']) : '-1';
     $suppliers_id = isset($_POST['suppliers_id']) ? intval($_POST['suppliers_id']) : '0';
 
     $goods_name_style = $_POST['goods_name_color'] . '+' . $_POST['goods_name_style'];
